Move logout and update password button to about activity
