+++ app/src/main/kotlin/io/github/projectbukirin/palaylab/About.kt
@@ -1,20 +1,53 @@
 package io.github.projectbukirin.palaylab
 
+import android.content.Intent
 import android.os.Bundle
 import android.text.method.LinkMovementMethod
 import android.view.View
+import android.widget.Button
 import android.widget.ImageButton
 import android.widget.TextView
+import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import com.google.firebase.auth.FirebaseAuth
 
 class About : AppCompatActivity() {
+    private lateinit var auth: FirebaseAuth
+
+    private lateinit var logoutBtn: Button
+    private lateinit var updatePass: Button
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        auth = FirebaseAuth.getInstance()
+
+        if(auth.currentUser == null){
+            val intent = Intent(this, LoginActivity::class.java)
+            startActivity(intent)
+            finish()
+        }else{
+        }
+
         setContentView(R.layout.activity_about)
         val imageButton = findViewById<View>(R.id.imageButton3) as ImageButton
         imageButton.setOnClickListener { backActivity() }
         val textView = findViewById<View>(R.id.link) as TextView
         textView.movementMethod = LinkMovementMethod.getInstance()
+
+        logoutBtn = findViewById(R.id.logout_btn)
+        updatePass = findViewById(R.id.update_pass_btn)
+
+        logoutBtn.setOnClickListener{
+            FirebaseAuth.getInstance().signOut()
+            val intent = Intent(this, LoginActivity::class.java)
+            startActivity(intent)
+            finish()
+        }
+
+        updatePass.setOnClickListener{
+            val intent = Intent(this, UpdatePassword::class.java)
+            startActivity(intent)
+        }
     }
 
     fun backActivity() {
