+++ app/src/main/res/layout/activity_main.xml
@@ -174,24 +174,6 @@
             android:id="@+id/textResult"
             android:layout_height="20dp" />
 
-        <Button
-            android:id="@+id/update_pass_btn"
-            android:layout_width="match_parent"
-            android:layout_height="56dp"
-            android:text="@string/update_pass"
-            app:layout_constraintBottom_toTopOf="@+id/logout_btn"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent" />
-
-        <Button
-            android:id="@+id/logout_btn"
-            android:layout_width="match_parent"
-            android:layout_height="56dp"
-            android:text="@string/logout"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent" />
-
     </TableLayout>
 
 </LinearLayout>
